libpayload: clear BSS, use 64b virtual_offset, remove debug output
Apparently BSS is not cleared by SeaBIOS, it needs to be cleared
manually.

+++ payloads/libpayload/arch/x86_64/virtual.c
@@ -29,7 +29,7 @@
 
 #include <unistd.h>
 
-unsigned long virtual_offset = 0;
+unsigned long long virtual_offset = 0;
 
 
 int getpagesize(void)
